pyproject.toml
[project]
name = "e-commerce"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asgiref==3.9.1",
    "black==25.9.0",
    "click==8.3.0",
    "django==5.2.5",
    "django-cleanup==9.0.0",
    "djangorestframework==3.16.1",
    "djangorestframework-simplejwt==5.5.1",
    "drf-yasg==1.21.10",
    "faker>=37.12.0",
    "flake8==7.3.0",
    "inflection==0.5.1",
    "iniconfig==2.3.0",
    "isort==7.0.0",
    "mccabe==0.7.0",
    "mypy-extensions==1.1.0",
    "packaging==25.0",
    "pathspec==0.12.1",
    "pillow==11.3.0",
    "platformdirs==4.5.0",
    "pluggy==1.6.0",
    "psycopg2-binary==2.9.11",
    "pycodestyle==2.14.0",
    "pyflakes==3.4.0",
    "pygments==2.19.2",
    "pyjwt==2.10.1",
    "pytest==8.4.2",
    "pytest-django==4.11.1",
    "python-dotenv==1.1.1",
    "pytokens==0.2.0",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "sqlparse==0.5.3",
    "uritemplate==4.2.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "root.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = """
    --strict-markers
    --disable-warnings
    --reuse-db
    --tb=short
    -v
"""
testpaths = ["Shop/tests", "tests"]
markers = [
    "slow: медленные тесты",
    "integration: интеграционные тесты",
    "unit: юнит-тесты"
]

[tool.isort]
profile = "black"
line_length = 79
multi_line_output = 3
include_trailing_comma = true
use_parentheses = true
skip = ["__init__.py"]
extend_skip = [".venv", "venv", "migrations"]

